Consolidate related routes into one file
